+++ src/app/reconcile/reconcile.ts
@@ -17,6 +17,7 @@ import { Transaction } from '../shared/transaction';
 import { Org } from '../shared/org';
 import { AppError } from '../shared/error';
 import { Util } from '../shared/util';
+import { DateUtil } from '../shared/dateutil';
 import { NgbModal, ModalDismissReasons } from '@ng-bootstrap/ng-bootstrap';
 import { ReconcileModal } from './reconcile-modal';
 import { Reconciliation } from './reconciliation';
@@ -87,8 +88,8 @@ export class ReconcilePage {
     let value = this.newReconcile.getRawValue();
 
     let rec = new Reconciliation();
-    rec.startDate = Util.getDateFromLocalDateString(value.startDate, this.org.timezone);
-    rec.endDate = Util.getDateFromLocalDateString(value.endDate, this.org.timezone);
+    rec.startDate = DateUtil.getDateFromLocalDateString(value.startDate, this.org.timezone);
+    rec.endDate = DateUtil.getDateFromLocalDateString(value.endDate, this.org.timezone);
     rec.startBalance = Math.round(parseFloat(value.startBalance) * Math.pow(10, this.account.precision));
     rec.endBalance = Math.round(parseFloat(value.endBalance) * Math.pow(10, this.account.precision));
 
@@ -106,7 +107,7 @@ export class ReconcilePage {
 
       this.newReconcile.patchValue(
         {
-          startDate: Util.getLocalDateString(rec.endDate, this.org.timezone),
+          startDate: DateUtil.getLocalDateString(rec.endDate, this.org.timezone),
           startBalance: rec.endBalance / Math.pow(10, this.account.precision),
           endBalance: 0,
           endDate: ''
@@ -203,7 +204,7 @@ export class ReconcilePage {
 
       if(!dates.length) {
         if(firstStartDate) {
-          this.newReconcile.patchValue({startDate: Util.getLocalDateString(firstStartDate, this.org.timezone)});
+          this.newReconcile.patchValue({startDate: DateUtil.getLocalDateString(firstStartDate, this.org.timezone)});
         }
         return;
       }
@@ -228,7 +229,7 @@ export class ReconcilePage {
 
       this.newReconcile.patchValue(
         {
-          startDate: Util.getLocalDateString(lastRec.endDate, this.org.timezone),
+          startDate: DateUtil.getLocalDateString(lastRec.endDate, this.org.timezone),
           startBalance: lastRec.endBalance / Math.pow(10, this.account.precision)
         }
       );
@@ -272,7 +273,7 @@ export class ReconcilePage {
         if(lastRec) {
           this.newReconcile.patchValue(
             {
-              startDate: Util.getLocalDateString(lastRec.endDate, this.org.timezone),
+              startDate: DateUtil.getLocalDateString(lastRec.endDate, this.org.timezone),
               startBalance: lastRec.endBalance / Math.pow(10, this.account.precision)
             }
           );
